+++ tests/testthat/test-align_move.R
@@ -13,16 +13,14 @@ test_that("align_move (default)", {
   x <- expect_is(align_move(m, res = "mean"), "MoveStack")
   expect_length(unique(unlist(move::timeLag(x, units = "secs"))), 1)
   
-  x <- expect_is(align_move(m, digit = "max"), "MoveStack")
-  expect_length(unique(unlist(move::timeLag(x, units = "secs"))), 1)
-  
-  x <- expect_is(align_move(m, digit = "mean"), "MoveStack")
-  expect_length(unique(unlist(move::timeLag(x, units = "secs"))), 1)
-  
   # false calls
   expect_error(align_move(NA)) # wrong class
   expect_error(align_move(m, res = FALSE))
-  expect_error(align_move(m, digit = FALSE))
   expect_error(align_move(m, unit = "abc"))
+  expect_error(align_move(move_data, res = 1, unit = "days"))
+  
+  # warnings
+  expect_warning(align_move(m, digit = "max"))
+  expect_warning(align_move(move_data, res = 13, unit = "hours"))
 })
 #}